requirements.txt
aiohttp
asyncpg
black
DuckDuck
fuzzywuzzy
googletrans==3.1.0a0
hikari
hikari-lightbulb
hikari-miru
hikari-toolbox
isort
Pillow
psutil
pyfiglet
python-dotenv
python-Levenshtein
uvloop
lightbulb-ext-filament